Change package structure a little
