+++ plugins/graphql/package.json
@@ -53,7 +53,6 @@
     "@graphql-codegen/add": "3.1.1",
     "@graphql-codegen/cli": "2.6.2",
     "@graphql-codegen/graphql-modules-preset": "2.3.8",
-    "@graphql-codegen/introspection": "2.1.1",
     "@graphql-codegen/schema-ast": "2.4.1",
     "@graphql-codegen/typescript": "2.4.8",
     "@graphql-codegen/typescript-resolvers": "2.6.1",
