+++ plugins/graphql/src/app/modules/component/component.graphql
