+++ plugins/graphql/src/app/modules/component/component.ts
@@ -4,7 +4,7 @@ import { createModule } from 'graphql-modules';
 
 export const Component = createModule({
   id: `component`,
-  typeDefs: loadFilesSync(resolvePackagePath('@internal/plugin-graphql', 'typedefs/component.graphql')),
+  typeDefs: loadFilesSync(resolvePackagePath('@internal/plugin-graphql', 'src/app/modules/component/component.graphql')),
   resolvers: {
     Lifecycle: {
       EXPERIMENTAL: 'experimental',
