+++ plugins/graphql/src/app/modules/domain/domain.graphql
