+++ plugins/graphql/src/app/modules/entity/entity.graphql
