+++ plugins/graphql/src/app/modules/entity/entity.ts
@@ -1,11 +1,11 @@
 import { resolvePackagePath } from '@backstage/backend-common';
 import { loadFilesSync } from '@graphql-tools/load-files';
 import { createModule } from 'graphql-modules';
-import { encodeId } from '../loaders';
+import { encodeId } from '../../loaders';
 
 export const Entity = createModule({
   id: 'entity',
-  typeDefs: loadFilesSync(resolvePackagePath('@internal/plugin-graphql', 'typedefs/entity.graphql')),
+  typeDefs: loadFilesSync(resolvePackagePath('@internal/plugin-graphql', 'src/app/modules/entity/entity.graphql')),
   resolvers: {
     Query: {
       entity: (
