+++ plugins/graphql/src/app/modules/location/location.graphql
