+++ plugins/graphql/src/app/modules/node/node.graphql
