+++ plugins/graphql/src/app/modules/node/node.ts
@@ -1,12 +1,12 @@
 import { resolvePackagePath } from '@backstage/backend-common';
 import { loadFilesSync } from '@graphql-tools/load-files';
 import { createModule } from 'graphql-modules';
-import { encodeId } from '../loaders';
-import { ResolverContext } from '../resolver-context';
+import { encodeId } from '../../loaders';
+import { ResolverContext } from '../../resolver-context';
 
 export const Node = createModule({
   id: 'node',
-  typeDefs: loadFilesSync(resolvePackagePath('@internal/plugin-graphql', 'typedefs/node.graphql')),
+  typeDefs: loadFilesSync(resolvePackagePath('@internal/plugin-graphql', 'src/app/modules/node/node.graphql')),
   resolvers: {
     Node: {
       id: async ({ id }: { id: string }, _: never, { loader }: ResolverContext): Promise<string | null> => {
