+++ plugins/graphql/src/app/modules/resource/resource.graphql
