+++ plugins/graphql/src/app/modules/system/system.graphql
