+++ plugins/graphql/src/app/modules/system/system.ts
@@ -2,7 +2,7 @@ import { resolvePackagePath } from '@backstage/backend-common';
 import { loadFilesSync } from '@graphql-tools/load-files';
 import { createModule } from 'graphql-modules';
 
-export const User = createModule({
-  id: `user`,
-  typeDefs: loadFilesSync(resolvePackagePath('@internal/plugin-graphql', 'typedefs/user.graphql')),
+export const System = createModule({
+  id: `system`,
+  typeDefs: loadFilesSync(resolvePackagePath('@internal/plugin-graphql', 'src/app/modules/system/system.graphql')),
 });
