+++ plugins/graphql/src/app/modules/template/template.graphql
