+++ plugins/graphql/src/app/modules/template/template.ts
@@ -5,7 +5,7 @@ import GraphQLJSON, { GraphQLJSONObject } from 'graphql-type-json';
 
 export const Template = createModule({
   id: 'Template',
-  typeDefs: loadFilesSync(resolvePackagePath('@internal/plugin-graphql', 'typedefs/template.graphql')),
+  typeDefs: loadFilesSync(resolvePackagePath('@internal/plugin-graphql', 'src/app/modules/template/template.graphql')),
   resolvers: {
     JSON: GraphQLJSON,
     JSONObject: GraphQLJSONObject,
