+++ plugins/graphql/src/app/modules/user/user.ts
@@ -2,7 +2,7 @@ import { resolvePackagePath } from '@backstage/backend-common';
 import { loadFilesSync } from '@graphql-tools/load-files';
 import { createModule } from 'graphql-modules';
 
-export const System = createModule({
-  id: `system`,
-  typeDefs: loadFilesSync(resolvePackagePath('@internal/plugin-graphql', 'typedefs/system.graphql')),
+export const User = createModule({
+  id: `user`,
+  typeDefs: loadFilesSync(resolvePackagePath('@internal/plugin-graphql', 'src/app/modules/user/user.graphql')),
 });
