+++ plugins/graphql/src/index.ts
@@ -14,4 +14,4 @@
  * limitations under the License.
  */
 
-export * from './service/router';
+export * from './app';
