+++ plugins/graphql/src/tests/setupTests.ts
