--- plugins/graphql/typedefs/shared.graphql
@@ -1,9 +1,0 @@
-enum Lifecycle {
-  EXPERIMENTAL
-  PRODUCTION
-  DEPRECATED
-}
-
-type Never implements Node {
-  id: ID!
-}
